refactor: separated clock classes into separate files and directories
