+++ packages/clock/src/Clock/index.ts
@@ -1,0 +1,1 @@
+export { Clock } from "./Clock"
