+++ packages/clock/src/Time/index.ts
@@ -1,0 +1,1 @@
+export { Time, throwIfTickRateIsZero } from "./Time"
