+++ packages/clock/src/utils/index.ts
@@ -1,0 +1,3 @@
+export const sleep = (ms: number) => {
+    return new Promise((resolve, reject) => setTimeout(resolve, ms));
+}
